Fix package entrypoints for TypeScript 7
