+++ types/ember/v2/test/function-ext.ts
@@ -1,9 +1,5 @@
 import Ember from "ember";
 
-declare global {
-    interface Function extends Ember.FunctionPrototypeExtensions {}
-}
-
 Ember.Object.extend({
     foo: "",
