+++ types/ember/v2/test/function-prototype-extensions.d.ts
@@ -1,0 +1,3 @@
+type EmberFunctionPrototypeExtensions = import("ember").default.FunctionPrototypeExtensions;
+
+interface Function extends EmberFunctionPrototypeExtensions {}
